[#156] Fix: 커스텀훅 및 초대받은 대시보드 오류 수정
